fix: prevent overwrite of relations results
If sentences contains more than one verb the result of the second verb
processing was overwriting the first one resulting with empty result as
it does not contain a subject. Now the processing stops after processing
the first verb.
To support finding results for several verbs in the same sentence this
must be changed.